Added a cast to make the type fall down to 32-bit when compiling for a 32-bit target without complaints.

+++ User Interfaces/Mac/View/LineGraph.m
@@ -62,7 +62,7 @@
 
 	// figure out the scale
 	CGFloat lowY = 2.0;
-	CGFloat highY = bounds.size.height - 2.0;
+	CGFloat highY = bounds.size.height - (CGFloat)2.0;
 	currentX = bounds.size.width;
 
 	// draw a digital value, thusly...
